Replace "use base" with "use parent"
As per the documentation, parent is encouraged these days:

"Unless you are using the fields pragma, consider this module
discouraged in favor of the lighter-weight parent."

And we're not using the fields pragma.

Since parent is only core from 5.10+ we add parent to the list of
dependencies for Perl 5.8 users.

Unlike base, parent actually fails when it can't load the base class
which means we can both remove some explicit use lines but also make the
codebase more robust since not all use lines were present, for example
Data::Buffer.